Move sighashtype FromStr tests to crypto
The EcdsaSighashType and TapSighashType FromStr tests are currently
located in bitcoin. This means testing regressions to the FromStr impl
requires running a significantly larger test set for no reason. These
should be located alongside the impls themselves in crypto.

Move sighashtype_fromstr_display tests to crypto::sighash.
